Made hail_hydra available to all redmew scenarios
