+++ config.lua
@@ -156,6 +156,31 @@ global.config = {
             },
         },
     },
+
+    -- spawns more units when one dies
+    hail_hydra = {
+        enabled = false,
+
+        -- any non-rounded number will turn into a chance to spawn an additional alien
+        -- example: 2.5 would spawn 2 for sure and 50% chance to spawn one additionally
+        hydras = {
+            -- spitters
+            ['small-spitter'] = {['small-worm-turret'] = 0.2},
+            ['medium-spitter'] = {['medium-worm-turret'] = 0.2},
+            ['big-spitter'] = {['big-worm-turret'] = 0.2},
+            ['behemoth-spitter'] = {['big-worm-turret'] = 0.4},
+
+            -- biters
+            ['medium-biter'] = {['small-biter'] = 1.2},
+            ['big-biter'] = {['medium-biter'] = 1.2},
+            ['behemoth-biter'] = {['big-biter'] = 1.2},
+
+            -- worms
+            ['small-worm-turret'] = {['small-biter'] = 2.5},
+            ['medium-worm-turret'] = {['small-biter'] = 2.5, ['medium-biter'] = 0.6},
+            ['big-worm-turret'] = {['small-biter'] = 3.8, ['medium-biter'] = 1.3, ['big-biter'] = 1.1},
+        },
+    },
 }
 
 return global.config
