+++ map_gen/Diggy/Feature/AlienSpawner.lua
@@ -46,22 +46,6 @@ end)
 
 local rocks_to_find = Template.diggy_rocks
 
-local function create_attack_command(position, target)
-    local command = {type = defines.command.attack_area, destination = position, radius = 10}
-    if target then
-        command = {
-            type = defines.command.compound,
-            structure_type = defines.compound_command.logical_or,
-            commands = {
-                {type = defines.command.attack, target = target},
-                command,
-            }
-        }
-    end
-
-    return command
-end
-
 ---Triggers mining at the collision_box of the alien, to free it
 local do_alien_mining = Token.register(function(params)
     local surface = params.surface
@@ -170,57 +154,8 @@ function AlienSpawner.register(config)
     local hail_hydra = config.hail_hydra
 
     if size(hail_hydra) > 0 then
-        Event.add(defines.events.on_entity_died, function (event)
-            local entity = event.entity
-            local name = entity.name
-
-            local hydras = hail_hydra[name]
-            if not hydras then
-                return
-            end
-
-            local position = entity.position
-            local force = entity.force
-            local evolution_factor = force.evolution_factor
-            local cause = event.cause
-
-            local surface = entity.surface
-            local create_entity = surface.create_entity
-            local find_non_colliding_position = surface.find_non_colliding_position
-
-            local command = create_attack_command(position, cause)
-
-            for hydra_spawn, amount in pairs(hydras) do
-                amount = amount + evolution_factor
-                local extra_chance = amount % 1
-                if extra_chance > 0 then
-                    if random() <= extra_chance then
-                        amount = ceil(amount)
-                    else
-                        amount = floor(amount)
-                    end
-                end
-                local particle_count
-
-                if amount > 4 then
-                    particle_count = 60
-                else
-                    particle_count = amount * 15
-                end
-
-                CreateParticles.blood_explosion(create_entity, particle_count, position)
-
-                for _ = amount, 1, -1  do
-                    position = find_non_colliding_position(hydra_spawn, position, 2, 0.4) or position
-                    local spawned = create_entity({name = hydra_spawn, force = force, position = position})
-                    if spawned and spawned.type == 'unit' then
-                        spawned.set_command(command)
-                    elseif spawned and cause then
-                        spawned.shooting_target = cause
-                    end
-                end
-            end
-        end)
+        global.config.hail_hydra.enabled = true
+        global.config.hail_hydra.hydras = hail_hydra
     end
 
     Event.add(Template.events.on_void_removed, function (event)
